Make notch/lead sample rate self-calibrating; root-cause and fix a real ~35% control-step throughput ceiling (double->float PID, -O0->-O2 build)
- Notch/lead filters no longer trust g_ctrl_rate_millihz (the requested
  throttle target) -- direct measurement found the real achieved
  control-step rate can differ from it by 12-73% depending on live Pi
  telemetry conditions, which vary a lot even within one session.
  Firmware now tracks a real inter-firing-interval EMA
  (g_measured_ctrl_interval_ms, reusing the PID's own dt_s) and
  recomputes both filters' coefficients from it every control step.
  Split notch_configure/lead_configure into *_compute_coeffs (safe to
  call every step) vs. *_reset_state (x1/x2/y1/y2 history, unchanged) --
  the combined originals would have zeroed the recursive filter state on
  every single control step, breaking them entirely. get_status gained
  meas_ctrl_rate_millihz= for direct visibility.
- Found and root-caused a real ~35% control-step throughput ceiling with
  axis2 enabled (555 raw I2C packets/s arriving, only ~355Hz driving a
  control step, zero checksum errors -- ruled out packet loss and
  confidence-filtering directly). Isolated to axis2 running
  PIDController.hpp's double-precision math a second time per packet on
  an MCU whose FPU is single-precision-only (double ops software-
  emulated) via a clean A/B (axis2 off closed ~30 points of the gap on
  its own).
- Converted PIDController.hpp, pid_wrapper.h/.cpp, and every main.c call
  site from double to float -- second explicit, user-approved deviation
  from Phil's class (first was dt-awareness). Same P/I/D terms, same
  anti-windup, same derivative filter, only the arithmetic type changed.
  .text shrank ~3.2KB, consistent with real double-emulation code removed.
- Found the project has been building at -O0 (debug, unoptimized) all
  session. Switched to -O2 (Debug/*/subdir.mk directly -- not persisted
  in .cproject, reapply by hand if ever regenerated from CubeIDE).
  .text shrank a further ~32%. Combined with the float fix, axis2-on
  unthrottled throughput went from ~64-66% of raw telemetry to ~90%+.
- Caught and fixed two real measurement-methodology bugs surfaced along
  the way: an impossible 10771 pkts/s reading traced to an unverified
  corrupted status read, and a systematic rate-inflation bug from using
  firmware's whole-second uptime= field (not host wall-clock) as a
  timing denominator for a short window.
- Real open item, not yet resolved: Kp=1.75/Ki=200 (historically always
  clean all session) now shows real ringing (33.8% overshoot) under the
  float PID, at an unchanged measured control rate -- not explained by
  a rate change; genuinely unclear whether it's a float-precision
  behavior difference or an unrelated physical rig change (a parallel
  flexure hardware redesign is also in progress on this same rig).
- CLAUDE.md: dated entry with the full investigation, root causes, and
  fixes. Slide deck: added the corrected notch Ki-ceiling comparison
  slide (fta_notch_ki_sweep_plot.py, new committed tool) with a
  correction note on the now-superseded prior slide.

+++ nucleo_firmware/camera_centroid_receiver/Core/Inc/pid_wrapper.h
@@ -38,23 +38,37 @@
  * still live-settable via pid_wrapper_set_ts) is now only the
  * construction-time FALLBACK for a dt<=0 call (e.g. the first call after
  * (re)construction) -- main.c's run_closed_loop_step measures real dt via
- * HAL_GetTick() and passes it to every pid_wrapper_calculate() call. */
+ * HAL_GetTick() and passes it to every pid_wrapper_calculate() call.
+ *
+ * 2026-08-27: switched from double to float throughout (this interface,
+ * PIDController.hpp, and every main.c call site) -- second, explicit,
+ * user-approved deviation from Phil's original class. This MCU's FPU
+ * (fpv4-sp-d16) is single-precision hardware only, so double math here
+ * was software-emulated; with the second axis (axis2) also running this
+ * same math per packet, direct hardware measurement found this cost a
+ * real ~35% of the achievable control-step rate (555 raw I2C packets/s
+ * arriving, only ~355Hz actually driving a control step) -- disabling
+ * axis2 alone (removing one of the two double-math calls per packet)
+ * closed most of that gap, isolating the cost to this math specifically.
+ * See PIDController.hpp's own docstring for the precision-risk assessment
+ * (low, for this pixel-scale, anti-windup-clamped, periodically-reset
+ * application). */
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void   pid_wrapper_init(double kp, double ki, double kd, double ts_s, double fc_hz,
-                         double out_min, double out_max);
-void   pid_wrapper_set_gains(double kp, double ki, double kd);
-void   pid_wrapper_set_fc(double fc_hz);  /* derivative low-pass cutoff, live-settable -- see
+void  pid_wrapper_init(float kp, float ki, float kd, float ts_s, float fc_hz,
+                        float out_min, float out_max);
+void  pid_wrapper_set_gains(float kp, float ki, float kd);
+void  pid_wrapper_set_fc(float fc_hz);  /* derivative low-pass cutoff, live-settable -- see
                                             * the g_fc_hz comment in pid_wrapper.cpp for why this
                                             * was added (20Hz default found too high relative to
                                             * this rig's ~15.3Hz resonance) */
-void   pid_wrapper_set_ts(double ts_s);  /* live-settable sample time, see pid_wrapper.cpp --
+void  pid_wrapper_set_ts(float ts_s);  /* live-settable sample time, see pid_wrapper.cpp --
                                             * added 2026-08-19 to test control-rate-throttling
                                             * hypotheses without a reflash per attempt */
-void   pid_wrapper_set_out_limits(double out_min, double out_max);  /* live-settable output
+void  pid_wrapper_set_out_limits(float out_min, float out_max);  /* live-settable output
                                             * saturation range passed straight to
                                             * PIDController::setOutputLimits() -- added 2026-08-19
                                             * to test whether tightening this (originally set to
@@ -63,16 +77,16 @@ void   pid_wrapper_set_out_limits(double out_min, double out_max);  /* live-sett
                                             * some of the settling-time/overshoot regression found
                                             * after adopting this class vs. the old hand-rolled
                                             * controller's proactive integral clamp -- see CLAUDE.md */
-double pid_wrapper_calculate(double setpoint_px, double measurement_px, double dt_s);
+float pid_wrapper_calculate(float setpoint_px, float measurement_px, float dt_s);
                                             /* dt_s: real elapsed seconds since the caller's own
                                             * last control step (<=0 to fall back to ts_) -- see
                                             * this header's own docstring above, 2026-08-19 */
-double pid_wrapper_calculate2(double setpoint_px, double measurement_px, double dt_s);
+float pid_wrapper_calculate2(float setpoint_px, float measurement_px, float dt_s);
                                             /* second axis (dac_x <- cy), identical Kp/Ki/Kd/ts_s/
                                             * fc_hz/limits to the first (reconstructed together by
                                             * every gain/fc/ts/limit setter) but its own independent
                                             * integral/derivative history -- added 2026-08-19 */
-void   pid_wrapper_reset(void);  /* resets BOTH axes' integral/derivative history together */
+void  pid_wrapper_reset(void);  /* resets BOTH axes' integral/derivative history together */
 
 #ifdef __cplusplus
 }
